Renamed StatsGui.java and tested Stats Activity
Working with Deny and we were having trouble when clicking the statsButton on the home screen. The crash only occurred on Deny's local repo, which was reporting up to date files. There is no crash on Robby's repo, so we're updating the master branch with Robby's repo so that everyone can access a working copy.

Also, GIt was giving a "pathspec" "did not match any files known to git" error, which researching reveals is a Windows case insensitivity error. Renamed the file to a temperorary intermediate name