Used Output.java, fixed finish() calls
I modified any Toast calls so that they would use the Output.toastMessage() method. This was an effort to cleanup code of onBackPressed() methods, which was using a backPressed counter that I deleted. (I'm making the assumption that when a user presses "back" once that they want to go back to the previous activity.

When buttons to new activities were being pressed, we were creating a new Intent and then calling finish() on the current activity. This deletes the activity from the stack, and so the user can't go back to the screen they were previously on. I fixed this by getting rid of all finish() calls, except when they were called in each activity's respective onBackPressed() method because the point of that method is to destroy the current activity.

Finally, I applied minimal coding standards so that code was more readable. It was mostly renaming variables and adding whitespace.