Added SQL database creation + insertions
Is now creating text files pertaining to the java classes for the objects to hold the column headings in a more organized way.

Also uses a new library to generate SQLite databases outside of android (apparently android comes bundled with one but I didn't want to extend this class with it for just one library).