Robotium Tests and Related Changes
First Robotium Test added that simple tests functioning of Ate activity. Can be used as a template for others probably. Also changed the build.gradle and androidmanifest to include the test runner and the target package. Which by the way is androidTest for UI tests. In orfer to run the test you will need to make a run configuration for the tests, using InstrumentedTestRunner, not JUnitTestRunner. and selecting the AndroidTest Java as the target package. Finally you will likely need to change your test artifact to Android Insturmentation Tests to run it, using the debug build variant.