Implemented food journal storage; database context retrieval
Removed Foods.java because DataFoods.java does the same thing.

Updated DatabaseDefinition to have a pointer to the last used DatabaseDefinition object so that the context (and thus th egetReadableDatabase) can be used easily with a static call to the DatabaseDefinition class (DatabaseDefinition.currentDatabase).
The last used object reference is stored.

The alternative strategy before this was to re-instantiate the database every time it was needed, which can be resource intensive. This technique allows the reference to be stored and re-usable.